Refactor card, card align, Home reponsiveness
