+++ src/App.js
@@ -13,16 +13,20 @@ import { PrivateRoute } from "./PrivateRote";
 import { Routes, Route, Link } from "react-router-dom";
 import { useAuth } from "./Contexts";
 import "./styles.css";
+import ShoppingCartIcon from "@material-ui/icons/ShoppingCart";
+import FavoriteIcon from "@material-ui/icons/Favorite";
 
 export function App() {
   const { currentUser } = useAuth();
 
   return (
-    <>
-      <div className="navbar">
-        <Link to="/" className="link">
-          <h1>NFT Baazar</h1>
-        </Link>
+    <div className="canvas">
+      <div className="nav">
+        <div class="logo">
+          <Link to="/" className="link">
+            <h1>NFT Baazar</h1>
+          </Link>
+        </div>
         <div className="nav-right">
           <ul className="nav-menu">
             <li>
@@ -32,13 +36,14 @@ export function App() {
               <Link to="/products">Products</Link>
             </li>
             <li>
-              <Link to="/cart">Cart</Link>
+              <Link to="/cart">
+                <ShoppingCartIcon />
+              </Link>
             </li>
             <li>
-              <Link to="/wishlist">WishList</Link>
-            </li>
-            <li>
-              <Link to="/test">Test</Link>
+              <Link to="/wishlist">
+                <FavoriteIcon />
+              </Link>
             </li>
             {!currentUser && (
               <li>
@@ -53,17 +58,19 @@ export function App() {
           </ul>
         </div>
       </div>
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/products" element={<ProductListing />} />
-        <PrivateRoute path="/cart" element={<Cart />} />
-        <PrivateRoute path="/wishlist" element={<WishList />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/signup" element={<Signup />} />
-        <Route path="/logout" element={<Logout />} />
-        <Route path="/products/:productId" element={<ProductDetails />} />
-        <Route path="/test" element={<TestAPI />} />
-      </Routes>
-    </>
+      <div className="main">
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/products" element={<ProductListing />} />
+          <PrivateRoute path="/cart" element={<Cart />} />
+          <PrivateRoute path="/wishlist" element={<WishList />} />
+          <Route path="/login" element={<Login />} />
+          <Route path="/signup" element={<Signup />} />
+          <Route path="/logout" element={<Logout />} />
+          <Route path="/products/:productId" element={<ProductDetails />} />
+          <Route path="/test" element={<TestAPI />} />
+        </Routes>
+      </div>
+    </div>
   );
 }
