+++ src/Pages/Home/Home.css
@@ -1,11 +1,10 @@
-
-
 .home {
     background-image: url(home_bg.png);
     background-repeat: no-repeat;
     background-size: 100vw 90vh;
     width: 100vw;
     height: 80vh;
+
 }
 
 .home__actions {
@@ -16,4 +15,25 @@
     flex-direction: column;
     align-items: center;
     
+}
+
+@media screen and (max-width: 600px) {
+    .home__actions {
+        position: absolute;
+        top:50%;
+        left: 50%;
+        width: 80vw;
+        transform: translate(-50%,-50%);
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        text-align: center;
+        background: white;
+        /* transform: rotate(-90deg); */
+    }
+
+    .home {
+        background-image: none;
+    }
+
 }
