+++ src/Pages/ProductDetails/ProductDetails.css
@@ -1,5 +1,5 @@
 .product-details {
-    margin: 2rem;
+    padding: 2rem;
     display: flex;
 }
