+++ src/Pages/index.js
@@ -1,5 +1,5 @@
 export { Cart } from "./Cart";
-export { ProductListing } from "./ProductListing";
+export { ProductListing } from "./ProductListing/ProductListing";
 export { WishList } from "./WishList";
 export { Login } from "./Login";
 export { Signup } from "./Signup";
